+++ src/components/model/pane.js
@@ -5,7 +5,7 @@
 
 import React, { Component } from 'react';
 import ModelTree from './tree';
-import Panes from '../panes';
+import Panes from '../tabbedpanes';
 import ModelNodeInspector from './nodeinspector';
 import ModelSelectionInspector from './selectioninspector';
 import StorageManager from '../../storagemanager';
@@ -62,7 +62,7 @@ export default class ModelPane extends Component {
 
 	render() {
 		if ( !this.props.editor ) {
-			return <div className="ck-inspector-panes__content__empty-wrapper">
+			return <div className="ck-inspector-tabbed-panes__content__empty-wrapper">
 				<p>Nothing to show. Attach another editor instance to start inspecting.</p>
 			</div>;
 		}
