+++ src/components/sidebar.css
@@ -11,7 +11,7 @@
 	display: flex;
 	width: 100%;
 
-	& > .ck-inspector-panes {
+	& > .ck-inspector-tabbed-panes {
 		min-width: var(--ck-inspector-explorer-width);
 		width: var(--ck-inspector-explorer-width);
 	}
