+++ src/components/ui.css
@@ -50,12 +50,12 @@
 	z-index: 9999;
 }
 
-.ck-inspector.ck-inspector_collapsed > .ck-inspector-panes > .ck-inspector-panes__navigation .ck-inspector-tabs {
+.ck-inspector.ck-inspector_collapsed > .ck-inspector-tabbed-panes > .ck-inspector-tabbed-panes__navigation .ck-inspector-tabs {
 	display: none;
 }
 
-.ck-inspector .ck-inspector-panes__navigation__logo,
-.ck-inspector .ck-inspector-panes__navigation__toggle {
+.ck-inspector .ck-inspector-tabbed-panes__navigation__logo,
+.ck-inspector .ck-inspector-tabbed-panes__navigation__toggle {
 	background-size: contain;
 	background-repeat: no-repeat;
 	background-position: center;
@@ -67,25 +67,25 @@
 	margin-right: 1em;
 }
 
-.ck-inspector .ck-inspector-panes__navigation__logo {
+.ck-inspector .ck-inspector-tabbed-panes__navigation__logo {
 	background-image: url("data:image/svg+xml;charset=UTF-8,%3csvg width='68' height='64' xmlns='http://www.w3.org/2000/svg'%3e%3cg fill='none' fill-rule='evenodd'%3e%3cpath d='M43.71 11.025a11.508 11.508 0 0 0-1.213 5.159c0 6.42 5.244 11.625 11.713 11.625.083 0 .167 0 .25-.002v16.282c0 1.955-1.05 3.761-2.756 4.739L30.986 60.7a5.548 5.548 0 0 1-5.512 0L4.756 48.828A5.464 5.464 0 0 1 2 44.089V20.344c0-1.955 1.05-3.76 2.756-4.738L25.474 3.733a5.548 5.548 0 0 1 5.512 0l12.724 7.292z' fill='%23FFF'/%3e%3cpath d='M45.684 8.79a12.604 12.604 0 0 0-1.329 5.65c0 7.032 5.744 12.733 12.829 12.733.091 0 .183-.001.274-.003v17.834c0 2.14-1.151 4.119-3.019 5.19L31.747 63.196a6.076 6.076 0 0 1-6.037 0L3.02 50.193A5.984 5.984 0 0 1 0 45.003V18.997c0-2.14 1.15-4.119 3.019-5.19L25.71.804a6.076 6.076 0 0 1 6.037 0l13.937 7.986zM16.244 20.68c-.834 0-1.51.671-1.51 1.498v.715c0 .828.676 1.498 1.51 1.498h25.489c.833 0 1.51-.67 1.51-1.498v-.715c0-.827-.677-1.498-1.51-1.498h-25.49zm0 9.227c-.834 0-1.51.671-1.51 1.498v.715c0 .828.676 1.498 1.51 1.498h18.479c.833 0 1.509-.67 1.509-1.498v-.715c0-.827-.676-1.498-1.51-1.498H16.244zm0 9.227c-.834 0-1.51.671-1.51 1.498v.715c0 .828.676 1.498 1.51 1.498h25.489c.833 0 1.51-.67 1.51-1.498v-.715c0-.827-.677-1.498-1.51-1.498h-25.49zm41.191-14.459c-5.835 0-10.565-4.695-10.565-10.486 0-5.792 4.73-10.487 10.565-10.487C63.27 3.703 68 8.398 68 14.19c0 5.791-4.73 10.486-10.565 10.486zm3.422-8.68c0-.467-.084-.875-.251-1.225a2.547 2.547 0 0 0-.686-.88 2.888 2.888 0 0 0-1.026-.531 4.418 4.418 0 0 0-1.259-.175c-.134 0-.283.006-.447.018a2.72 2.72 0 0 0-.446.07l.075-1.4h3.587v-1.8h-5.462l-.214 5.06c.319-.116.682-.21 1.089-.28.406-.071.77-.107 1.088-.107.218 0 .437.021.655.063.218.041.413.114.585.218s.313.244.422.419c.109.175.163.391.163.65 0 .424-.132.745-.396.961a1.434 1.434 0 0 1-.938.325c-.352 0-.656-.1-.912-.3-.256-.2-.43-.453-.523-.762l-1.925.588c.1.35.258.664.472.943.214.279.47.514.767.706.298.191.63.339.995.443.365.104.749.156 1.151.156.437 0 .86-.064 1.272-.193.41-.13.778-.323 1.1-.581.324-.258.582-.585.775-.981.193-.396.29-.864.29-1.405z' fill='%231EBC61' fill-rule='nonzero'/%3e%3c/g%3e%3c/svg%3e");
 	width: 1.8em;
 	height: 1.8em;
 	margin-left: 1em;
 }
 
-.ck-inspector .ck-inspector-panes__navigation__toggle {
+.ck-inspector .ck-inspector-tabbed-panes__navigation__toggle {
 	background-image: url(data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTYiIGhlaWdodD0iMTYiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHBhdGggZD0iTTIgNS41bDUuMjkzIDUuMjkzYTEgMSAwIDAgMCAxLjQxNCAwTDE0IDUuNSIgZmlsbC1ydWxlPSJub256ZXJvIiBzdHJva2U9IiM3NDc0NzQiIHN0cm9rZS13aWR0aD0iMiIgZmlsbD0ibm9uZSIgc3Ryb2tlLWxpbmVjYXA9InJvdW5kIiBzdHJva2UtbGluZWpvaW49InJvdW5kIi8+PC9zdmc+);
 	width: 1.6em;
 	height: 1.6em;
 	opacity: .7;
 }
 
-.ck-inspector .ck-inspector-panes__navigation__toggle:hover {
+.ck-inspector .ck-inspector-tabbed-panes__navigation__toggle:hover {
 	opacity: 1;
 }
 
-.ck-inspector .ck-inspector-panes__navigation__toggle.ck-inspector-panes__navigation__toggle_up {
+.ck-inspector .ck-inspector-tabbed-panes__navigation__toggle.ck-inspector-tabbed-panes__navigation__toggle_up {
 	transform: rotate(180deg);
 }
