+++ src/components/view/tree.js
@@ -49,8 +49,8 @@ class ViewTree extends Component {
 		const tree = this.getEditorViewTree();
 
 		return <div className="ck-inspector__document-tree">
-			<div className="ck-inspector-panes">
-				<div className="ck-inspector-panes__navigation">
+			<div className="ck-inspector-tabbed-panes">
+				<div className="ck-inspector-tabbed-panes__navigation">
 					<div className="ck-inspector__document-tree__config">
 						<Select
 							id="view-root-select"
@@ -69,7 +69,7 @@ class ViewTree extends Component {
 						/>
 					</div>
 				</div>
-				<div className="ck-inspector-panes__content">
+				<div className="ck-inspector-tabbed-panes__content">
 					<Tree
 						items={tree}
 						onClick={this.props.onClick}
